Refactor cli root command creation to make it vendorable and extensible * Extract all root command construction to commands package and export NewDockerCommand * Extract daemon files to its own package
